Renderable redesigned
