+++ CS300/CS300.vcxproj
@@ -253,7 +253,7 @@
     <ClInclude Include="Graphics\Primitives\Model.h" />
     <ClInclude Include="Graphics\Primitives\Texture.h" />
     <ClInclude Include="Object.h" />
-    <ClInclude Include="Renderables.h" />
+    <ClInclude Include="Graphics\Primitives\Renderables.h" />
     <ClInclude Include="Core\ECSystem\Scene.h" />
     <ClInclude Include="Core\Window\SDLWindow.h" />
     <ClInclude Include="Dependencies\GLEW\Include\eglew.h" />
