+++ CS300/Core/ECSystem/Scene.h
@@ -13,7 +13,7 @@
 #include <algorithm>
 #include <execution>
 #include "CS300Parser.h"
-#include "Renderables.h"
+#include "Graphics/Primitives/Renderables.h"
 
 namespace Core {
 	class Renderable;
